+++ NEWS.md
@@ -12,6 +12,7 @@ features of `procimg()`.
 * added the argument `labels.stack` to `plot.rspec`, which allows the use of 
 custom spectra labels in stacked plots.
 * all deprecated functions and arguments have now been fully removed.
+* `as.rspec()` now accepts both numeric and character vectors to identify the wavelength column.  
 * numerous under-the-hood changes for stability and speed, with thanks to 
 three reviewers and an associate editor at MEE.